Now generate SUPER MEGA HARD sudoku
The program randomly pop like 80 of the 81 values of the sudokus.
If popping one would result in several solutions then it's  we put the value
back.
When a random cell is chosen to be popped, we register the fact it should not
be attempted to be popped again. This way we eventually treat all the cells
of the sudoku.
However the fact making the easiest sudokus is the hardest to code is kind of ironic